Adapt custom routes helper to zeitwerk conventions
In order to have the routes splitted into api and webui
routes, a couple of helper classes got created and added
to the eager_load_paths in the past. The folder that contains
those helper classes, contains the route definitions as well.
This has the side effect, that zeitwerk derives constants
from those (so 'ApiRoutes' and 'WebuiRoutes'). But there
is no module/class that corresponds to those constant names.

Also the custom routes helper defining multiple classes
in one file which doesn't work with Zeitwerk anymore (it doesen't
break the codebase, since the file that contains the declartions
was still required manually in the routes.rb).

In order to fix it, the custom routes helper got moved to the
'app/lib' folder, which gets eager_loaded/auto_loaded
by default and classes/files got splitted accordingly.
