+++ src/api/app/lib/routes_helper/routes_constraints.rb
@@ -1,0 +1,25 @@
+module RoutesHelper
+  module RoutesConstraints
+    CONS = {
+      arch: %r{[^/]*},
+      binary: %r{[^/]*},
+      filename: %r{[^/]*},
+      id: %r{\d*},
+      login: %r{[^/]*},
+      package: %r{[^/]*},
+      package_name: %r{[^/]*},
+      project: %r{[^/]*},
+      project_name: %r{[^/]*},
+      maintained_project: %r{[^/]*},
+      repository: %r{[^/]*},
+      repository_name: %r{[^/]*},
+      service: %r{\w[^/]*},
+      title: %r{[^/]*},
+      user: %r{[^/]*},
+      repository_publish_build_id: %r{[^/]*},
+      workflow_project: %r{[^/]*},
+      staging_project_name: %r{[^/]*},
+      staging_project_copy_name: %r{[^/]*}
+    }.freeze
+  end
+end
