+++ src/api/app/lib/routes_helper/webui_matcher.rb
@@ -1,0 +1,12 @@
+module RoutesHelper
+  class WebuiMatcher
+    class InvalidRequestFormat < APIError
+    end
+
+    def self.matches?(request)
+      request.format.to_sym != :xml
+    rescue ArgumentError => e
+      raise InvalidRequestFormat, e.to_s
+    end
+  end
+end
