+++ src/api/config/application.rb
@@ -19,7 +19,6 @@ module OBSApi
 
     # Custom directories with classes and modules you want to be autoloadable.
     # config.autoload_paths += %W(#{config.root}/extras)
-    config.eager_load_paths << Rails.root.join('config', 'routes') if Rails.env.development?
 
     # Only load the plugins named here, in the order given (default is alphabetical).
     # :all can be used as a placeholder for all plugins not explicitly named.
