+++ src/api/config/initializers/zeitwerk.rb
@@ -1,6 +1,7 @@
 Rails.autoloaders.each do |autoloader|
   autoloader.inflector = Zeitwerk::Inflector.new
   autoloader.inflector.inflect(
+    'api_matcher' => 'APIMatcher',
     'cve_parser' => 'CVEParser',
     'meta_xml_validator' => 'MetaXMLValidator',
     'obs_quality_categories_finder' => 'OBSQualityCategoriesFinder',
