+++ src/api/config/routes.rb
@@ -1,5 +1,3 @@
-require_relative 'routes/routes_helper'
-
 class ActionDispatch::Routing::Mapper
   def draw(routes_name)
     instance_eval(File.read(Rails.root.join("config/routes/#{routes_name}_routes.rb")))
